Simple Drag and Drop
Highest Number educational game
OLE-Nepal is a non-governmental organization dedicated to assisting the
Government of Nepal in meeting its Education for All goals by developing
freely accessible, open-source Information and Communication. Most of
their activity before used flash, but now they are slowly transitioning
into htlf 5 because flash uses more memory and most of the device that
OLE-Nepal uses are limited to its computer power.
In their program for the activities they use code that are efficient and
easy. For animation they use sprite image, which are a collection of
images put into a single image. A web page with many images can take a
long time to load and generates multiple server requests.Some of the
advantages of using an sprite images are, it reduces the number of
server request and save bandwidth.
Also, for image file they use png file due to its advantages. Some of
its advantages are: It is best suited for serving in,internet as per
it’s name,Portable Network Graphics, Smaller in size, Supports
transparency.
And as for sound file, OGG are best because  it is common on most of the
browser and it is the smallest size among the other sound file format.
For this project, I have used
Jquery, css, html5 and javascript